% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDRRNG.R
\name{HDRUniformGen}
\alias{HDRUniformGen}
\title{HDR Uniform RNG}
\usage{
HDRUniformGen(n, Var = sample(1:1e+08, 1), Ent = 1000, Att1 = 0, Att2 = 0)
}
\arguments{
\item{n}{The number of trials to be generated}

\item{Var}{Unique #ID for the simulation between 1 and 100,000,000}

\item{Ent}{A secondary optional unique ID between 1 and 100,000,000; defaults to 1000}

\item{Att1}{A tertiary optional attribute ID between 1 and 100,000,000; defaults to 0}

\item{Att2}{A quaternary optional attribute ID between 1 and 100,000,000; defaults to 0}
}
\value{
A vector n length with psuedorandom uniform values between 0 and 1
}
\description{
Creates a vector of uniform values ranging between 0 and 1 to be used in
conjunction with quantile functions to generate repeatable simulations.
Allows for the use of two values, "Var" and "Ent" to create repeatable
simulations across platforms.
}
\examples{
randoms <- HDRUniformGen(100000)
qnorm(randoms, 10, 2)
qpois(randoms, 10)
qexp(randoms, 2)
qgamma(randoms, 5)
qlnorm(randoms, 2, .5)
}
