% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPExport.R
\name{SLURPExportCSV}
\alias{SLURPExportCSV}
\title{Export Data as .CSV SLURP}
\usage{
SLURPExportCSV(
  dataframe,
  filename = "",
  slurpname = "",
  origin = "",
  meta = NULL
)
}
\arguments{
\item{dataframe}{The dataframe containing the SIPs to export as a SLURP.}

\item{filename}{A string containing the desired file name. Add ".csv" to the
string, otherwise the file will write in a basic file format used by the write()
function.}

\item{slurpname}{An optional string containing the desired name of the SLURP}

\item{origin}{An optional string containing the an identifier for who made the
SLURP.}

\item{meta}{An optional dataframe containing metadata to include in export. See
the SLURPMetaDF function example for the correct format.}
}
\value{
A message indicating a success.
}
\description{
Exports a dataframe and optional metadata as a SIP Library in .csv format to
the current wd, allowing a user to export SLURPs generated in R.
}
\examples{
SIPdf <- data.frame(matrix(ncol = 5, nrow = 2000))
colnames(SIPdf) <- c("simulate1","simulate2","simulate3","simulate4","simulate5")
for(i in 1:5) {
 SIPdf[,i] <- runif(2000)
 SIPdf[,i] <- ifelse(SIPdf[,i] > .5, 0, 50000)
}

SLURPExportXML(SIPdf, testslurp.xml, average = TRUE, median = TRUE)
}
