% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPExport.R
\name{SLURPExportXML}
\alias{SLURPExportXML}
\title{Export Data as .XML SLURP}
\usage{
SLURPExportXML(
  dataframe,
  filename,
  index = FALSE,
  provenance = "",
  csvr = 4,
  average = FALSE,
  median = FALSE,
  meta = NULL
)
}
\arguments{
\item{dataframe}{The dataframe containing the SIP(s).}

\item{filename}{Name for the exported .xml.}

\item{index}{Logical, optionally increase index by 1}

\item{provenance}{Optional string to indicate origin}

\item{csvr}{The number digits to the right of the decimal to include}

\item{average}{Logical, set to TRUE to include the averages as metadata}

\item{median}{Logical, set to TRUE to include the medians as metadata}

\item{meta}{Optional dataframe created from SLURPMetaDF function. A user may also
specify dataframe of metadata where the column name matches the name of the SIP,
and the final vector contains the names of the metadata. See the example for
SLURPMetaDF function.}
}
\description{
Exports a dataframe and optional metadata as a SIP Library in .xml format to
the current wd, allowing a user to export SLURPs generated in R.
}
\examples{
SIPdf <- data.frame(matrix(ncol = 5, nrow = 2000))
colnames(SIPdf) <- c("simulate1","simulate2","simulate3","simulate4","simulate5")
for(i in 1:5) {
 SIPdf[,i] <- runif(2000)
 SIPdf[,i] <- ifelse(SIPdf[,i] > .5, 0, 50000)
}

SLURPExportXML(SIPdf, testslurp.xml, average = TRUE, median = TRUE)
}
