% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPExport.R
\name{SLURPMetaDF}
\alias{SLURPMetaDF}
\title{Create Metadata}
\usage{
SLURPMetaDF(df, IDcol, IDvect, metanamesvect)
}
\arguments{
\item{df}{A dataframe containing data you wish to turn into metadata.}

\item{IDcol}{Name of column with ID's}

\item{IDvect}{A vector containing the ID's associated with the SIP(s).}

\item{metanamesvect}{A vector containing the column names of the metadata.}
}
\value{
A dataframe with metadata to be used in association with the
"ExportSIPLib" SIP export function.
}
\description{
Returns data frame containing metadata about the specified SIP(s). Useful for
constructing metadata from a structured dataset to be included with a SIP library.
}
\examples{
data <- mtcars
data$ID <- rownames(mtcars)
meta <- SLURPMetaDF(data, "ID", rownames(mtcars)[1:4], colnames(mtcars)[1:4])
}
